Refactor VirtualKeyContainer to hold a dictionary instead of an array
To improve lookup performance, `VirtualKeyContainer` is now backed by
a dictionary instead of an array. This means that lookup will happen
in constant time because keys are computed when mapping the current
input source. Memory usage might be slightly bigger and mapping
might take slightly longer, but seeing as index usually happens
when the input changes, having faster lookup takes precedence.

This can be improved upon in the future.